+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Date format : day/month/year
 
 ## 1.9.0 - ??/??/2022 - Foundry v10 Compatibility
 - Updated the System to FoundryVTT v10.
+- Removed restriction on technique types on drop (Sheet and 20Q).
 - Added a `game.user.isFirstGM` property for some traitements (socket and migration) to prevent multiple executions with multiple GM connected.
 
 ## 1.8.2 - 24/06/2022 - Unofficial Italian translation